+++ .gitignore
@@ -1,12 +1,11 @@
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
-##
-## Get latest from https://github.com/github/gitignore/blob/main/VisualStudio.gitignore
 
+# =========== PROJECT BUILD FOLDERS ===========
 /bin-int
 /bin
+/build
 
-# User-specific files
 *.rsuser
 *.suo
 *.user
@@ -15,18 +14,39 @@
 *.sln
 .vs/
 
-/* Aggiungi queste righe per escludere specifici file di progetto di Visual Studio */
 *.vcxproj
 *.vcxproj.filters
 *.vcxproj.user
 
-# User-specific files (MonoDevelop/Xamarin Studio)
+# =========== LINUX / GCC / MAKE ===========
+*.o
+*.d
+
+*.a
+*.so
+*.so.*
+*.dylib
+
+Makefile
+*.make
+
+*.out
+*.app
+
+# =========== VISUAL STUDIO CODE ===========
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+*.code-workspace
+
+# =========== STANDARD VISUAL STUDIO ===========
+
 *.userprefs
 
-# Mono auto generated files
 mono_crash.*
 
-# Build results
 [Dd]ebug/
 [Dd]ebugPublic/
 [Rr]elease/
@@ -42,43 +62,31 @@ bld/
 [Ll]og/
 [Ll]ogs/
 
-# Visual Studio 2015/2017 cache/options directory
 .vs/
-# Uncomment if you have tasks that create the project's static files in wwwroot
-#wwwroot/
 
-# Visual Studio 2017 auto generated files
 Generated\ Files/
 
-# MSTest test Results
 [Tt]est[Rr]esult*/
 [Bb]uild[Ll]og.*
 
-# NUnit
 *.VisualState.xml
 TestResult.xml
 nunit-*.xml
 
-# Build Results of an ATL Project
 [Dd]ebugPS/
 [Rr]eleasePS/
 dlldata.c
 
-# Benchmark Results
 BenchmarkDotNet.Artifacts/
 
-# .NET Core
 project.lock.json
 project.fragment.lock.json
 artifacts/
 
-# ASP.NET Scaffolding
 ScaffoldingReadMe.txt
 
-# StyleCop
 StyleCopReport.xml
 
-# Files built by Visual Studio
 *_i.c
 *_p.c
 *_h.h
@@ -108,11 +116,8 @@ StyleCopReport.xml
 *.svclog
 *.scc
 
-
-# Chutzpah Test files
 _Chutzpah*
 
-# Visual C++ cache files
 ipch/
 *.aps
 *.ncb
@@ -123,61 +128,46 @@ ipch/
 *.VC.db
 *.VC.VC.opendb
 
-# Visual Studio profiler
 *.psess
 *.vsp
 *.vspx
 *.sap
 
-# Visual Studio Trace Files
 *.e2e
 
-# TFS 2012 Local Workspace
 $tf/
 
-# Guidance Automation Toolkit
 *.gpState
 
-# ReSharper is a .NET coding add-in
 _ReSharper*/
 *.[Rr]e[Ss]harper
 *.DotSettings.user
 
-# TeamCity is a build add-in
 _TeamCity*
 
-# DotCover is a Code Coverage Tool
 *.dotCover
 
-# AxoCover is a Code Coverage Tool
 .axoCover/*
 !.axoCover/settings.json
 
-# Coverlet is a free, cross platform Code Coverage Tool
 coverage*.json
 coverage*.xml
 coverage*.info
 
-# Visual Studio code coverage results
 *.coverage
 *.coveragexml
 
-# NCrunch
 _NCrunch_*
 .*crunch*.local.xml
 nCrunchTemp_*
 
-# MightyMoose
 *.mm.*
 AutoTest.Net/
 
-# Web workbench (sass)
 .sass-cache/
 
-# Installshield output folder
 [Ee]xpress/
 
-# DocProject is a documentation generator add-in
 DocProject/buildhelp/
 DocProject/Help/*.HxT
 DocProject/Help/*.HxC
@@ -187,45 +177,29 @@ DocProject/Help/*.hhp
 DocProject/Help/Html2
 DocProject/Help/html
 
-# Click-Once directory
 publish/
 
-# Publish Web Output
 *.[Pp]ublish.xml
 *.azurePubxml
-# Note: Comment the next line if you want to checkin your web deploy settings,
-# but database connection strings (with potential passwords) will be unencrypted
 *.pubxml
 *.publishproj
 
-# Microsoft Azure Web App publish settings. Comment the next line if you want to
-# checkin your Azure Web App publish settings, but sensitive information contained
-# in these scripts will be unencrypted
 PublishScripts/
 
-# NuGet Packages
 *.nupkg
-# NuGet Symbol Packages
 *.snupkg
-# The packages folder can be ignored because of Package Restore
 **/[Pp]ackages/*
-# except build/, which is used as an MSBuild target.
 !**/[Pp]ackages/build/
-# Uncomment if necessary however generally it will be regenerated when needed
-#!**/[Pp]ackages/repositories.config
-# NuGet v3's project.json files produces more ignorable files
 *.nuget.props
 *.nuget.targets
 
-# Microsoft Azure Build Output
 csx/
 *.build.csdef
 
-# Microsoft Azure Emulator
 ecf/
 rcf/
 
-# Windows Store app package directories and files
+
 AppPackages/
 BundleArtifacts/
 Package.StoreAssociation.xml
@@ -234,10 +208,8 @@ _pkginfo.txt
 *.appxbundle
 *.appxupload
 
-# Visual Studio cache files
-# files ending in .cache can be ignored
+
 *.[Cc]ache
-# but keep track of directories ending in .cache
 !?*.[Cc]ache/
 
 # Others
@@ -250,5 +222,3 @@ ClientBin/
 *.pfx
 *.publishsettings
 orleans.codegen.cs
-
-# Including strong name files can present a security
