+++ .vscode/launch.json
@@ -2,15 +2,37 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Sandbox (Debug)", 
-            "type": "cppvsdbg",
+            // --- LINUX CONFIGURATION ---
+            "name": "Sandbox (Linux Debug)",
+            "type": "cppdbg",
             "request": "launch",
-            "program": "${workspaceFolder}/bin/Debug-windows-x86_64/Sandbox/Sandbox.exe", 
+            "program": "${workspaceFolder}/bin/Debug-linux-x86_64/Sandbox/Sandbox",
             "args": [],
             "stopAtEntry": false,
-            "cwd": "${workspaceFolder}/Sandbox", 
+            "cwd": "${workspaceFolder}/Sandbox",
+            "environment": [],
+            "externalConsole": false,
+            "MIMode": "gdb",
+            "setupCommands": [
+                {
+                    "description": "Enable pretty-printing for gdb",
+                    "text": "-enable-pretty-printing",
+                    "ignoreFailures": true
+                }
+            ],
+            "preLaunchTask": "Build TrEngine"
+        },
+        {
+            // --- WINDOWS CONFIGURATION ---
+            "name": "Sandbox (Windows Debug)",
+            "type": "cppvsdbg",
+            "request": "launch",
+            "program": "${workspaceFolder}/bin/Debug-windows-x86_64/Sandbox/Sandbox.exe",
+            "args": [],
+            "stopAtEntry": false,
+            "cwd": "${workspaceFolder}/Sandbox",
             "console": "externalTerminal",
-            "preLaunchTask": "Build TrEngine (Debug)"
+            "preLaunchTask": "Build TrEngine"
         }
     ]
 }
