+++ .vscode/tasks.json
@@ -2,20 +2,29 @@
     "version": "2.0.0",
     "tasks": [
         {
-            "label": "Build TrEngine (Debug)",
+            "label": "Build TrEngine",
             "type": "shell",
-            "command": ".\\Build.bat",
-            "args": [
-                "TrEngine.sln",
-                "/p:Configuration=Debug",
-                "/p:Platform=x64",
-                "/t:Build"
-            ],
             "group": {
                 "kind": "build",
                 "isDefault": true
             },
-            "problemMatcher": "$msCompile"
+            // --- Windows ---
+            "windows": {
+                "command": ".\\Build.bat",
+                "args": [
+                    "TrEngine.sln",
+                    "/p:Configuration=Debug",
+                    "/p:Platform=x64",
+                    "/t:Build"
+                ],
+                "problemMatcher": "$msCompile"
+            },
+            // --- Linux ---
+            "linux": {
+                "command": "./Build.sh",
+                "args": [],
+                "problemMatcher": "$gcc"
+            }
         }
     ]
 }
