+++ TrEngine/src/TrEngine.h
@@ -1,12 +1,12 @@
 #pragma once
-//For use by TrEngine applications
+// For use by TrEngine applications
 #include "TrEngine/Application.h"
 
 #include "TrEngine/Input.h"
 #include "TrEngine/KeyCode.h"
 #include "TrEngine/MouseButtonCode.h"
-#include "TrEngine/event/ImGuiEventHandler.h"
-#include "TrEngine/event/Event.h"
+#include "TrEngine/Event/ImGuiEventHandler.h"
+#include "TrEngine/Event/Event.h"
 
 #include "TrEngine/Layer.h"
 #include "TrEngine/ImGui/ImGuiLayer.h"
@@ -25,4 +25,3 @@
 //------------------ Entry Point -------------------
 #include "TrEngine/EntryPoint.h"
 //--------------------------------------------------
-
