+++ TrEngine/src/TrEngine/EntryPoint.h
@@ -6,13 +6,32 @@ extern TrEngine::Application *TrEngine::CreateApplication();
 
 int main(int argc, char **argv)
 {
-	TrEngine::Log::Init();
-	TE_CORE_WARN("Inizialized log!");
-	int a = 5;
-	TE_INFO("Hello! Var={0}", a);
+    TrEngine::Log::Init();
+    TE_CORE_WARN("Initialized Log (Windows)!");
 
-	auto app = TrEngine::CreateApplication();
-	app->Run();
-	delete app;
+    int a = 5;
+    TE_INFO("Hello! Var={0}", a);
+
+    auto app = TrEngine::CreateApplication();
+    app->Run();
+    delete app;
 }
-#endif // TE_PLATFORM_WINDOWS
+
+#elif defined(TE_PLATFORM_LINUX)
+
+extern TrEngine::Application *TrEngine::CreateApplication();
+
+int main(int argc, char **argv)
+{
+    TrEngine::Log::Init();
+    TE_CORE_WARN("Initialized Log (Linux)!");
+
+    auto app = TrEngine::CreateApplication();
+    app->Run();
+    delete app;
+    return 0; // Linux want return
+}
+
+#else
+#error TrEngine support Windows and Linux only!
+#endif
