+++ TrEngine/src/TrEngine/Event/ApplicationEvent.cpp
@@ -1,0 +1,11 @@
+#include "Tepch.h"
+#include "ApplicationEvent.h"
+
+TrEngine::WindowCloseEvent::~WindowCloseEvent() = default;
+
+TrEngine::AppTickEvent::~AppTickEvent() = default;
+
+TrEngine::AppUpdateEvent::~AppUpdateEvent() = default;
+
+TrEngine::AppRenderEvent::~AppRenderEvent() = default;
+
