+++ TrEngine/src/TrEngine/Platform/OpenGL/OpenGLContext.h
@@ -11,12 +11,12 @@ namespace TrEngine
 		OpenGLContext() = default;
 		~OpenGLContext();
 
-		explicit OpenGLContext(GLFWwindow* windowHandle);
+		explicit OpenGLContext(GLFWwindow *windowHandle);
 
 		void Init() override;
 		void SwapBuffers() override;
+
 	private:
-		GLFWwindow* m_WindowHandle;
+		GLFWwindow *m_WindowHandle;
 	};
 }
-
