+++ TrEngine/src/TrEngine/Window.h
@@ -4,7 +4,7 @@
 
 #include "TrEngine/Core.h"
 
-#include "TrEngine/event/Event.h"
+#include "TrEngine/Event/Event.h"
 
 namespace TrEngine
 {
