--- TrEngine/src/TrEngine/event/ApplicationEvent.cpp
@@ -1,11 +1,0 @@
-#include "Tepch.h"
-#include "ApplicationEvent.h"
-
-TrEngine::WindowCloseEvent::~WindowCloseEvent() = default;
-
-TrEngine::AppTickEvent::~AppTickEvent() = default;
-
-TrEngine::AppUpdateEvent::~AppUpdateEvent() = default;
-
-TrEngine::AppRenderEvent::~AppRenderEvent() = default;
-
